requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
python-dotenv
openai
colorama